refactor(to_vtk): Refactor in response to comments
